Spliting code in modules
