+++ src/routes/health_check.rs
@@ -1,1 +1,6 @@
+use rocket::http::Status;
 
+#[get("/health_check")]
+pub async fn health_check() -> Status {
+    Status::Ok
+}
