+++ src/routes/mod.rs
@@ -1,5 +1,5 @@
-mod health_check;
-mod subscriptions;
+pub mod health_check;
+pub mod subscriptions;
 
 pub use health_check::*;
 pub use subscriptions::*;
